+++ Nasubi/Model/Trending/Trending.MediaItem.swift
@@ -27,9 +27,9 @@ extension Trending {
       
       var genreIds: [GenreId]?
 
-      var popularity: Float?
+      var popularity: Double?
       var voteCount: Int?
-      var voteAverage: Float?
+      var voteAverage: Double?
       
       var posterPath: String?
       var backdropPath: String?
